hello_exercise/requirements.txt
Django==1.11.6
djangorestframework==3.6.4
Pygments==2.2.0